requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
sqlalchemy>=2.0
requests>=2.31
requests-oauthlib>=1.3
python-dotenv>=1.0
